+++ test/device_test.rb
@@ -1,6 +1,15 @@
 require_relative "test_helper"
 
 class DeviceTest < Minitest::Test
+  def test_type
+    assert_equal Torch.device("cpu:0").type, "cpu"
+  end
+
+  def test_index
+    assert_nil Torch.device("cpu").index
+    assert_equal 0, Torch.device("cpu:0").index
+  end
+
   def test_equal
     a = Torch.device("cpu")
     b = Torch.device("cpu")